Split models file into multiple files
